requirements.txt
torch
torchvision
diffusers
transformers
einops
pandas
Pillow
accelerate
